requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
exceptiongroup==1.0.4
idna==3.4
iniconfig==1.1.1
ndg-httpsclient==0.5.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.5
pusher==3.3.0
pyasn1==0.4.8
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytest==7.2.0
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
